requirements.txt
django
djangorestframework
python-decouple
psycopg2-binary
django-filter
drf-yasg
django-rest-swagger